Improve use of coordinates and enable to put a dash in non-working days
Change the script to use the input coordinates to find out which days are
working days. The actual coordinates to write in the PS document are
computed with hardcoded offsets (the original script already used some
hardcoded information).

By default, the script will put a dot in the signature for working days.
The user can use the '--dashed' argument to put a dash in the
signature of non-working days instead of the dot.